feat: redirect logged-in users from root
